@@ -37,19 +37,19 @@ THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
 #include "FreeSurround/_KissFFTGuts.h"
-/* The guts header contains all the multiplication and addition macros that are
- defined for
- fixed or floating point complex numbers.  It also delares the kf_ internal
- functions.
+#include <random>
+#include <vector>
+
+/* The guts header contains all the multiplication and addition macros
+   that are defined for fixed or floating point complex numbers.
+   It also declares the kf_ internal functions.
  */
 
-static void kf_bfly2(kiss_fft_cpx *Fout, const size_t fstride,
-                     const kiss_fft_cfg st, int m) {
-  kiss_fft_cpx *Fout2;
-  kiss_fft_cpx *tw1 = st->twiddles;
-  kiss_fft_cpx t;
-  Fout2 = Fout + m;
+static void kf_bfly2(kiss_fft_cpx *Fout, const size_t fstride, const kiss_fft_cfg st, int m) {
+  const kiss_fft_cpx *tw1 = st->twiddles;
+  kiss_fft_cpx *Fout2 = Fout + m;
   do {
+    kiss_fft_cpx t;
     C_FIXDIV(*Fout, 2);
     C_FIXDIV(*Fout2, 2);
 
@@ -62,17 +62,15 @@ static void kf_bfly2(kiss_fft_cpx *Fout, const size_t fstride,
   } while (--m);
 }
 
-static void kf_bfly4(kiss_fft_cpx *Fout, const size_t fstride,
-                     const kiss_fft_cfg st, const size_t m) {
+static void kf_bfly4(kiss_fft_cpx *Fout, const size_t fstride, const kiss_fft_cfg st, const size_t m) {
   kiss_fft_cpx *tw1, *tw2, *tw3;
-  kiss_fft_cpx scratch[6];
   size_t k = m;
-  const size_t m2 = 2 * m;
-  const size_t m3 = 3 * m;
+  const size_t m2 = 2 * m, m3 = 3 * m;
 
   tw3 = tw2 = tw1 = st->twiddles;
 
   do {
+    kiss_fft_cpx scratch[6];
     C_FIXDIV(*Fout, 4);
     C_FIXDIV(Fout[m], 4);
     C_FIXDIV(Fout[m2], 4);
@@ -107,18 +105,16 @@ static void kf_bfly4(kiss_fft_cpx *Fout, const size_t fstride,
   } while (--k);
 }
 
-static void kf_bfly3(kiss_fft_cpx *Fout, const size_t fstride,
-                     const kiss_fft_cfg st, size_t m) {
+static void kf_bfly3(kiss_fft_cpx *Fout, const size_t fstride, const kiss_fft_cfg st, const size_t m) {
   size_t k = m;
   const size_t m2 = 2 * m;
   kiss_fft_cpx *tw1, *tw2;
-  kiss_fft_cpx scratch[5];
-  kiss_fft_cpx epi3;
-  epi3 = st->twiddles[fstride * m];
+  const auto [r, i] = st->twiddles[fstride * m];
 
   tw1 = tw2 = st->twiddles;
 
   do {
+    kiss_fft_cpx scratch[5];
     C_FIXDIV(*Fout, 3);
     C_FIXDIV(Fout[m], 3);
     C_FIXDIV(Fout[m2], 3);
@@ -134,7 +130,7 @@ static void kf_bfly3(kiss_fft_cpx *Fout, const size_t fstride,
     Fout[m].r = Fout->r - HALF_OF(scratch[3].r);
     Fout[m].i = Fout->i - HALF_OF(scratch[3].i);
 
-    C_MULBYSCALAR(scratch[0], epi3.i);
+    C_MULBYSCALAR(scratch[0], i);
 
     C_ADDTO(*Fout, scratch[3]);
 
@@ -148,25 +144,14 @@ static void kf_bfly3(kiss_fft_cpx *Fout, const size_t fstride,
   } while (--k);
 }
 
-static void kf_bfly5(kiss_fft_cpx *Fout, const size_t fstride,
-                     const kiss_fft_cfg st, int m) {
-  kiss_fft_cpx *Fout0, *Fout1, *Fout2, *Fout3, *Fout4;
-  int u;
+static void kf_bfly5(kiss_fft_cpx *Fout, const size_t fstride, const kiss_fft_cfg st, const int m) {
+  kiss_fft_cpx *Fout0 = Fout, *Fout1 = Fout0 + m, *Fout2 = Fout0 + 2 * m, *Fout3 = Fout0 + 3 * m,
+      *Fout4 = Fout0 + 4 * m;
   kiss_fft_cpx scratch[13];
-  kiss_fft_cpx *twiddles = st->twiddles;
-  kiss_fft_cpx *tw;
-  kiss_fft_cpx ya, yb;
-  ya = twiddles[fstride * m];
-  yb = twiddles[fstride * 2 * m];
+  const kiss_fft_cpx *twiddles = st->twiddles, *tw = st->twiddles, ya = twiddles[fstride * m],
+      yb = twiddles[fstride * 2 * m];
 
-  Fout0 = Fout;
-  Fout1 = Fout0 + m;
-  Fout2 = Fout0 + 2 * m;
-  Fout3 = Fout0 + 3 * m;
-  Fout4 = Fout0 + 4 * m;
-
-  tw = st->twiddles;
-  for (u = 0; u < m; ++u) {
+  for (int u = 0; u < m; ++u) {
     C_FIXDIV(*Fout0, 5);
     C_FIXDIV(*Fout1, 5);
     C_FIXDIV(*Fout2, 5);
@@ -187,10 +172,8 @@ static void kf_bfly5(kiss_fft_cpx *Fout, const size_t fstride,
     Fout0->r += scratch[7].r + scratch[8].r;
     Fout0->i += scratch[7].i + scratch[8].i;
 
-    scratch[5].r =
-        scratch[0].r + S_MUL(scratch[7].r, ya.r) + S_MUL(scratch[8].r, yb.r);
-    scratch[5].i =
-        scratch[0].i + S_MUL(scratch[7].i, ya.r) + S_MUL(scratch[8].i, yb.r);
+    scratch[5].r = scratch[0].r + S_MUL(scratch[7].r, ya.r) + S_MUL(scratch[8].r, yb.r);
+    scratch[5].i = scratch[0].i + S_MUL(scratch[7].i, ya.r) + S_MUL(scratch[8].i, yb.r);
 
     scratch[6].r = S_MUL(scratch[10].i, ya.i) + S_MUL(scratch[9].i, yb.i);
     scratch[6].i = -S_MUL(scratch[10].r, ya.i) - S_MUL(scratch[9].r, yb.i);
@@ -198,10 +181,8 @@ static void kf_bfly5(kiss_fft_cpx *Fout, const size_t fstride,
     C_SUB(*Fout1, scratch[5], scratch[6]);
     C_ADD(*Fout4, scratch[5], scratch[6]);
 
-    scratch[11].r =
-        scratch[0].r + S_MUL(scratch[7].r, yb.r) + S_MUL(scratch[8].r, ya.r);
-    scratch[11].i =
-        scratch[0].i + S_MUL(scratch[7].i, yb.r) + S_MUL(scratch[8].i, ya.r);
+    scratch[11].r = scratch[0].r + S_MUL(scratch[7].r, yb.r) + S_MUL(scratch[8].r, ya.r);
+    scratch[11].i = scratch[0].i + S_MUL(scratch[7].i, yb.r) + S_MUL(scratch[8].i, ya.r);
     scratch[12].r = -S_MUL(scratch[10].i, yb.i) + S_MUL(scratch[9].i, ya.i);
     scratch[12].i = S_MUL(scratch[10].r, yb.i) - S_MUL(scratch[9].r, ya.i);
 
@@ -217,43 +198,35 @@ static void kf_bfly5(kiss_fft_cpx *Fout, const size_t fstride,
 }
 
 /* perform the butterfly for one stage of a mixed radix FFT */
-static void kf_bfly_generic(kiss_fft_cpx *Fout, const size_t fstride,
-                            const kiss_fft_cfg st, int m, int p) {
-  int u, k, q1, q;
-  kiss_fft_cpx *twiddles = st->twiddles;
-  kiss_fft_cpx t;
-  int Norig = st->nfft;
+static void kf_bfly_generic(kiss_fft_cpx *Fout, const size_t fstride, const kiss_fft_cfg st, const int m, const int p) {
+  const kiss_fft_cpx *twiddles = st->twiddles;
+  const int Norig = st->nfft;
 
-  kiss_fft_cpx *scratch =
-      (kiss_fft_cpx *)KISS_FFT_TMP_ALLOC(sizeof(kiss_fft_cpx) * p);
+  const auto scratch = static_cast<kiss_fft_cpx *>(KISS_FFT_TMP_ALLOC(sizeof(kiss_fft_cpx) * p));
 
-  for (u = 0; u < m; ++u) {
-    k = u;
-    for (q1 = 0; q1 < p; ++q1) {
-      scratch[q1] = Fout[k];
+  for (int u = 0; u < m; ++u) {
+    for (int q1 = 0, i = u; q1 < p; ++q1, i += m) {
+      scratch[q1] = Fout[i];
       C_FIXDIV(scratch[q1], p);
-      k += m;
     }
 
-    k = u;
-    for (q1 = 0; q1 < p; ++q1) {
+    for (int q1 = 0, j = u; q1 < p; ++q1, j += m) {
       int twidx = 0;
-      Fout[k] = scratch[0];
-      for (q = 1; q < p; ++q) {
-        twidx += static_cast<int>(fstride) * k;
+      Fout[j] = scratch[0];
+      for (int q = 1; q < p; ++q) {
+        kiss_fft_cpx t;
+        twidx += static_cast<int>(fstride) * j;
         if (twidx >= Norig)
           twidx -= Norig;
         C_MUL(t, scratch[q], twiddles[twidx]);
-        C_ADDTO(Fout[k], t);
+        C_ADDTO(Fout[j], t);
       }
-      k += m;
     }
   }
   KISS_FFT_TMP_FREE(scratch);
 }
 
-static void kf_work(kiss_fft_cpx *Fout, const kiss_fft_cpx *f,
-                    const size_t fstride, int in_stride, int *factors,
+static void kf_work(kiss_fft_cpx *Fout, const kiss_fft_cpx *f, const size_t fstride, int in_stride, int *factors,
                     const kiss_fft_cfg st) {
   kiss_fft_cpx *Fout_beg = Fout;
   const int p = *factors++; /* the radix  */
@@ -269,26 +242,25 @@ static void kf_work(kiss_fft_cpx *Fout, const kiss_fft_cpx *f,
 // execute the p different work units in different threads
 #pragma omp parallel for
     for (k = 0; k < p; ++k)
-      kf_work(Fout + k * m, f + fstride * in_stride * k, fstride * p, in_stride,
-              factors, st);
+      kf_work(Fout + k * m, f + fstride * in_stride * k, fstride * p, in_stride, factors, st);
     // all threads have joined by this point
 
     switch (p) {
-    case 2:
-      kf_bfly2(Fout, fstride, st, m);
-      break;
-    case 3:
-      kf_bfly3(Fout, fstride, st, m);
-      break;
-    case 4:
-      kf_bfly4(Fout, fstride, st, m);
-      break;
-    case 5:
-      kf_bfly5(Fout, fstride, st, m);
-      break;
-    default:
-      kf_bfly_generic(Fout, fstride, st, m, p);
-      break;
+      case 2:
+        kf_bfly2(Fout, fstride, st, m);
+        break;
+      case 3:
+        kf_bfly3(Fout, fstride, st, m);
+        break;
+      case 4:
+        kf_bfly4(Fout, fstride, st, m);
+        break;
+      case 5:
+        kf_bfly5(Fout, fstride, st, m);
+        break;
+      default:
+        kf_bfly_generic(Fout, fstride, st, m, p);
+        break;
     }
     return;
   }
@@ -314,54 +286,89 @@ static void kf_work(kiss_fft_cpx *Fout, const kiss_fft_cpx *f,
 
   // recombine the p smaller DFTs
   switch (p) {
-  case 2:
-    kf_bfly2(Fout, fstride, st, m);
-    break;
-  case 3:
-    kf_bfly3(Fout, fstride, st, m);
-    break;
-  case 4:
-    kf_bfly4(Fout, fstride, st, m);
-    break;
-  case 5:
-    kf_bfly5(Fout, fstride, st, m);
-    break;
-  default:
-    kf_bfly_generic(Fout, fstride, st, m, p);
-    break;
+    case 2:
+      kf_bfly2(Fout, fstride, st, m);
+      break;
+    case 3:
+      kf_bfly3(Fout, fstride, st, m);
+      break;
+    case 4:
+      kf_bfly4(Fout, fstride, st, m);
+      break;
+    case 5:
+      kf_bfly5(Fout, fstride, st, m);
+      break;
+    default:
+      kf_bfly_generic(Fout, fstride, st, m, p);
+      break;
   }
 }
 
-/*  facbuf is populated by p1,m1,p2,m2, ...
-    where
-    p[i] * m[i] = m[i-1]
-    m0 = n                  */
-static void kf_factor(int n, int *facbuf) {
-  int p = 4;
-  double floor_sqrt;
-  floor_sqrt = floor(sqrt((double)n));
+/**
+ * @brief Implements Pollard's Rho algorithm to find a non-trivial factor of n.
+ *
+ * This function uses Pollard's Rho algorithm, a probabilistic integer factorization method.
+ * The algorithm generates a sequence of numbers based on a quadratic recurrence and uses the
+ * "tortoise and hare" technique to detect cycles. If a cycle is detected, the GCD of the
+ * difference between two sequence values and n is computed, which yields a factor of n.
+ *
+ * If n is even, the function immediately returns 2. Otherwise, the algorithm iterates
+ * until it finds a non-trivial factor or determines that no factor exists.
+ *
+ * @param n The integer to factor. The function will return a non-trivial factor of n.
+ *
+ * @return A non-trivial factor of n if found, otherwise 0 if no factor is found.
+ */
+int pollards_rho(const int n) {
+  if (n % 2 == 0)
+    return 2;
 
-  /*factor out powers of 4, powers of 2, then any remaining primes */
-  do {
-    while (n % p) {
-      switch (p) {
-      case 4:
-        p = 2;
-        break;
-      case 2:
-        p = 3;
-        break;
-      default:
-        p += 2;
-        break;
-      }
-      if (p > floor_sqrt)
-        p = n; /* no more factors, skip to end */
+  std::random_device rd;
+  std::mt19937 gen(rd());
+  std::uniform_int_distribution dist(1, n - 1);
+
+  int64_t x = dist(gen);
+  int64_t y = x;
+  const int64_t c = dist(gen);
+  int64_t d = 1;
+
+  while (d == 1) {
+    x = (x * x + c) % n;
+    y = (y * y + c) % n;
+    y = (y * y + c) % n;
+    d = std::gcd(std::llabs(x - y), static_cast<int64_t>(n));
+  }
+
+  return d == n ? 0 : static_cast<int>(d);
+}
+
+/**
+ * @brief Factorizes a number using Pollard's Rho algorithm.
+ *
+ * This function attempts to factor the given number `n` by repeatedly calling the Pollard's Rho algorithm.
+ * It continuously divides `n` by each found factor and stores the factors in the provided `facbuf` array.
+ * This method returns the prime factors of `n` and updates the array with factor pairs:
+ * each factor and the quotient of `n` after dividing by the factor.
+ *
+ * The factorization process continues until all factors of `n` are found,
+ * or until Pollard's Rho fails to find further factors.
+ *
+ * @param n The integer to factor. It will be reduced during the process.
+ * @param facbuf A pointer to an array where the factors will be stored.
+ * The factors are stored as pairs: each factor and the quotient of n after dividing by the factor.
+ */
+void kf_factor(int n, int *facbuf) {
+  while (n > 1) {
+    const int factor = pollards_rho(n);
+    if (factor == 0)
+      break;
+
+    while (n % factor == 0) {
+      n /= factor;
+      *facbuf++ = factor;
+      *facbuf++ = n;
     }
-    n /= p;
-    *facbuf++ = p;
-    *facbuf++ = n;
-  } while (n > 1);
+  }
 }
 
 /*
@@ -372,73 +379,87 @@ static void kf_factor(int n, int *facbuf) {
  * such,
  * It can be freed with free(), rather than a kiss_fft-specific function.
  * */
-kiss_fft_cfg kiss_fft_alloc(int nfft, int inverse_fft, void *mem,
-                            size_t *lenmem) {
-  kiss_fft_cfg st = NULL;
-  size_t memneeded = sizeof(struct kiss_fft_state) +
-                     sizeof(kiss_fft_cpx) * (nfft - 1); /* twiddle factors*/
+kiss_fft_cfg kiss_fft_alloc(const int nfft, const int inverse_fft, void *mem, size_t *lenmem) {
+  kiss_fft_cfg st = nullptr;
+  const size_t memneeded = sizeof(struct kiss_fft_state) + sizeof(kiss_fft_cpx) * (nfft - 1); /* twiddle factors*/
 
-  if (lenmem == NULL) {
-    st = (kiss_fft_cfg) new char[memneeded];
+  if (lenmem == nullptr) {
+    st = reinterpret_cast<kiss_fft_cfg>(new char[memneeded]);
   } else {
-    if (mem != NULL && *lenmem >= memneeded)
-      st = (kiss_fft_cfg)mem;
+    if (mem != nullptr && *lenmem >= memneeded)
+      st = static_cast<kiss_fft_cfg>(mem);
     *lenmem = memneeded;
   }
-  if (st) {
-    int i;
-    st->nfft = nfft;
-    st->inverse = inverse_fft;
-
-    for (i = 0; i < nfft; ++i) {
-      const double pi =
-          3.141592653589793238462643383279502884197169399375105820974944;
-      double phase = -2 * pi * i / nfft;
-      if (st->inverse)
-        phase *= -1;
-      kf_cexp(st->twiddles + i, phase);
-    }
-
-    kf_factor(nfft, st->factors);
+  if (!st) {
+    return st;
   }
+  st->nfft = nfft;
+  st->inverse = inverse_fft;
+
+  for (int i = 0; i < nfft; ++i) {
+    double phase = -2 * pi * i / nfft;
+    if (st->inverse)
+      phase *= -1;
+    kf_cexp(st->twiddles + i, phase);
+  }
+
+  kf_factor(nfft, st->factors);
   return st;
 }
 
-void kiss_fft_stride(kiss_fft_cfg st, const kiss_fft_cpx *fin,
-                     kiss_fft_cpx *fout, int in_stride) {
-  if (fin == fout) {
-    // NOTE: this is not really an in-place FFT algorithm.
-    // It just performs an out-of-place FFT into a temp buffer
-    kiss_fft_cpx *tmpbuf =
-        (kiss_fft_cpx *)KISS_FFT_TMP_ALLOC(sizeof(kiss_fft_cpx) * st->nfft);
-    kf_work(tmpbuf, fin, 1, in_stride, st->factors, st);
-    memcpy(fout, tmpbuf, sizeof(kiss_fft_cpx) * st->nfft);
-    KISS_FFT_TMP_FREE(tmpbuf);
-  } else {
-    kf_work(fout, fin, 1, in_stride, st->factors, st);
+void kiss_fft_stride(const kiss_fft_cfg cfg, const kiss_fft_cpx *fin, kiss_fft_cpx *fout, const int fin_stride) {
+  if (fin != fout) {
+    kf_work(fout, fin, 1, fin_stride, cfg->factors, cfg);
+    return;
   }
+  // NOTE: this is not really an in-place FFT algorithm.
+  // It just performs an out-of-place FFT into a temp buffer
+  auto *tmpbuf = static_cast<kiss_fft_cpx *>(KISS_FFT_TMP_ALLOC(sizeof(kiss_fft_cpx) * cfg->nfft));
+  kf_work(tmpbuf, fin, 1, fin_stride, cfg->factors, cfg);
+  memcpy(fout, tmpbuf, sizeof(kiss_fft_cpx) * cfg->nfft);
+  KISS_FFT_TMP_FREE(tmpbuf);
 }
 
-void kiss_fft(kiss_fft_cfg cfg, const kiss_fft_cpx *fin, kiss_fft_cpx *fout) {
+void kiss_fft(const kiss_fft_cfg cfg, const kiss_fft_cpx *fin, kiss_fft_cpx *fout) {
   kiss_fft_stride(cfg, fin, fout, 1);
 }
 
-void kiss_fft_cleanup(void) {
-  // nothing needed any more
-}
+/**
+ * Finds the next largest integer that can be expressed as a product of
+ * the prime factors 2, 3, and 5. This ensures the number is factorable
+ * by these primes, making it suitable for optimized FFT computations.
+ *
+ * @param n The starting integer to search from.
+ * @return The smallest integer greater than or equal to `n`
+ *         divisible only by the primes 2, 3, and 5.
+ */
+int kiss_fft_next_fast_size(const int n) {
+  std::vector hammingNumbers = {1}; // Start with 1 as the smallest Hamming number
+  int i2 = 0, i3 = 0, i5 = 0; // Pointers for multiples of 2, 3, and 5
 
-int kiss_fft_next_fast_size(int n) {
-  while (1) {
-    int m = n;
-    while ((m % 2) == 0)
-      m /= 2;
-    while ((m % 3) == 0)
-      m /= 3;
-    while ((m % 5) == 0)
-      m /= 5;
-    if (m <= 1)
-      break; /* n is completely factorable by twos, threes, and fives */
-    n++;
+  while (true) {
+    // Generate the next candidates by multiplying with 2, 3, and 5
+    int next2 = hammingNumbers[i2] * 2;
+    int next3 = hammingNumbers[i3] * 3;
+    int next5 = hammingNumbers[i5] * 5;
+
+    // Find the smallest candidate
+    int nextHamming = std::min({next2, next3, next5});
+
+    // If the candidate is >= n, return it
+    if (nextHamming >= n) {
+      return nextHamming;
+    }
+
+    // Add the smallest candidate to the list
+    hammingNumbers.push_back(nextHamming);
+
+    // Increment the respective pointer(s)
+    if (nextHamming == next2)
+      i2++;
+    if (nextHamming == next3)
+      i3++;
+    if (nextHamming == next5)
+      i5++;
   }
-  return n;
 }
