@@ -38,6 +38,7 @@ THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "FreeSurround/KissFFTR.h"
 #include "FreeSurround/_KissFFTGuts.h"
+#include <cstdio>
 
 struct kiss_fftr_state {
   kiss_fft_cfg substate;
@@ -48,40 +49,36 @@ struct kiss_fftr_state {
 #endif
 };
 
-kiss_fftr_cfg kiss_fftr_alloc(int nfft, int inverse_fft, void *mem,
-                              size_t *lenmem) {
-  int i;
-  kiss_fftr_cfg st = NULL;
+kiss_fftr_cfg kiss_fftr_alloc(int nfft, const int inverse_fft, void *mem, size_t *lenmem) {
+  kiss_fftr_cfg st = nullptr;
   size_t subsize = 65536 * 4, memneeded = 0;
 
   if (nfft & 1) {
     fprintf(stderr, "Real FFT optimization must be even.\n");
-    return NULL;
+    return nullptr;
   }
   nfft >>= 1;
 
-  kiss_fft_alloc(nfft, inverse_fft, NULL, &subsize);
-  memneeded = sizeof(struct kiss_fftr_state) + subsize +
-              sizeof(kiss_fft_cpx) * (nfft * 3 / 2);
+  kiss_fft_alloc(nfft, inverse_fft, nullptr, &subsize);
+  memneeded = sizeof(kiss_fftr_state) + subsize + sizeof(kiss_fft_cpx) * (nfft * 3 / 2);
 
-  if (lenmem == NULL) {
-    st = (kiss_fftr_cfg)malloc(memneeded);
+  if (lenmem == nullptr) {
+    st = static_cast<kiss_fftr_cfg>(malloc(memneeded));
   } else {
     if (*lenmem >= memneeded)
-      st = (kiss_fftr_cfg)mem;
+      st = static_cast<kiss_fftr_cfg>(mem);
     *lenmem = memneeded;
   }
   if (!st)
-    return NULL;
+    return nullptr;
 
-  st->substate = (kiss_fft_cfg)(st + 1); /*just beyond kiss_fftr_state struct */
-  st->tmpbuf = (kiss_fft_cpx *)(((char *)st->substate) + subsize);
+  st->substate = reinterpret_cast<kiss_fft_cfg>(st + 1); /*just beyond kiss_fftr_state struct */
+  st->tmpbuf = reinterpret_cast<kiss_fft_cpx *>(reinterpret_cast<char *>(st->substate) + subsize);
   st->super_twiddles = st->tmpbuf + nfft;
   kiss_fft_alloc(nfft, inverse_fft, st->substate, &subsize);
 
-  for (i = 0; i < nfft / 2; ++i) {
-    double phase =
-        -3.14159265358979323846264338327 * ((double)(i + 1) / nfft + .5);
+  for (int i = 0; i < nfft / 2; ++i) {
+    double phase = -pi * (static_cast<double>(i + 1) / nfft + .5);
     if (inverse_fft)
       phase *= -1;
     kf_cexp(st->super_twiddles + i, phase);
@@ -89,21 +86,19 @@ kiss_fftr_cfg kiss_fftr_alloc(int nfft, int inverse_fft, void *mem,
   return st;
 }
 
-void kiss_fftr(kiss_fftr_cfg st, const kiss_fft_scalar *timedata,
-               kiss_fft_cpx *freqdata) {
+void kiss_fftr(kiss_fftr_cfg cfg, const kiss_fft_scalar *timedata, kiss_fft_cpx *freqdata) {
   /* input buffer timedata is stored row-wise */
-  int k, ncfft;
-  kiss_fft_cpx fpnk, fpk, f1k, f2k, tw, tdc;
+  kiss_fft_cpx tdc;
 
-  if (st->substate->inverse) {
+  if (cfg->substate->inverse) {
     fprintf(stderr, "kiss fft usage error: improper alloc\n");
     exit(1);
   }
 
-  ncfft = st->substate->nfft;
+  int ncfft = cfg->substate->nfft;
 
   /*perform the parallel fft of two real signals packed in real,imag*/
-  kiss_fft(st->substate, (const kiss_fft_cpx *)timedata, st->tmpbuf);
+  kiss_fft(cfg->substate, reinterpret_cast<const kiss_fft_cpx *>(timedata), cfg->tmpbuf);
   /* The real part of the DC element of the frequency spectrum in st->tmpbuf
    * contains the sum of the even-numbered elements of the input time sequence
    * The imag part is the sum of the odd-numbered elements
@@ -115,29 +110,30 @@ void kiss_fftr(kiss_fftr_cfg st, const kiss_fft_scalar *timedata,
    *      yielding Nyquist bin of input time sequence
    */
 
-  tdc.r = st->tmpbuf[0].r;
-  tdc.i = st->tmpbuf[0].i;
+  tdc.r = cfg->tmpbuf[0].r;
+  tdc.i = cfg->tmpbuf[0].i;
   C_FIXDIV(tdc, 2);
   CHECK_OVERFLOW_OP(tdc.r, +, tdc.i);
   CHECK_OVERFLOW_OP(tdc.r, -, tdc.i);
   freqdata[0].r = tdc.r + tdc.i;
   freqdata[ncfft].r = tdc.r - tdc.i;
 #ifdef USE_SIMD
-  freqdata[ncfft].i = freqdata[0].i = _mm_set1_ps(0);
+    freqdata[ncfft].i = freqdata[0].i = _mm_set1_ps(0);
 #else
   freqdata[ncfft].i = freqdata[0].i = 0;
 #endif
 
-  for (k = 1; k <= ncfft / 2; ++k) {
-    fpk = st->tmpbuf[k];
-    fpnk.r = st->tmpbuf[ncfft - k].r;
-    fpnk.i = -st->tmpbuf[ncfft - k].i;
+  for (int k = 1; k <= ncfft / 2; ++k) {
+    kiss_fft_cpx fpnk, f1k, f2k, tw;
+    const kiss_fft_cpx fpk = cfg->tmpbuf[k];
+    fpnk.r = cfg->tmpbuf[ncfft - k].r;
+    fpnk.i = -cfg->tmpbuf[ncfft - k].i;
     C_FIXDIV(fpk, 2);
     C_FIXDIV(fpnk, 2);
 
     C_ADD(f1k, fpk, fpnk);
     C_SUB(f2k, fpk, fpnk);
-    C_MUL(tw, f2k, st->super_twiddles[k - 1]);
+    C_MUL(tw, f2k, cfg->super_twiddles[k - 1]);
 
     freqdata[k].r = HALF_OF(f1k.r + tw.r);
     freqdata[k].i = HALF_OF(f1k.i + tw.i);
@@ -146,25 +142,23 @@ void kiss_fftr(kiss_fftr_cfg st, const kiss_fft_scalar *timedata,
   }
 }
 
-void kiss_fftri(kiss_fftr_cfg st, const kiss_fft_cpx *freqdata,
-                kiss_fft_scalar *timedata) {
+void kiss_fftri(kiss_fftr_cfg cfg, const kiss_fft_cpx *freqdata, kiss_fft_scalar *timedata) {
   /* input buffer timedata is stored row-wise */
-  int k, ncfft;
 
-  if (st->substate->inverse == 0) {
+  if (cfg->substate->inverse == 0) {
     fprintf(stderr, "kiss fft usage error: improper alloc\n");
     exit(1);
   }
 
-  ncfft = st->substate->nfft;
+  int ncfft = cfg->substate->nfft;
 
-  st->tmpbuf[0].r = freqdata[0].r + freqdata[ncfft].r;
-  st->tmpbuf[0].i = freqdata[0].r - freqdata[ncfft].r;
+  cfg->tmpbuf[0].r = freqdata[0].r + freqdata[ncfft].r;
+  cfg->tmpbuf[0].i = freqdata[0].r - freqdata[ncfft].r;
   C_FIXDIV(st->tmpbuf[0], 2);
 
-  for (k = 1; k <= ncfft / 2; ++k) {
-    kiss_fft_cpx fk, fnkc, fek, fok, tmp;
-    fk = freqdata[k];
+  for (int k = 1; k <= ncfft / 2; ++k) {
+    kiss_fft_cpx fnkc, fek, fok, tmp;
+    const kiss_fft_cpx fk = freqdata[k];
     fnkc.r = freqdata[ncfft - k].r;
     fnkc.i = -freqdata[ncfft - k].i;
     C_FIXDIV(fk, 2);
@@ -172,14 +166,14 @@ void kiss_fftri(kiss_fftr_cfg st, const kiss_fft_cpx *freqdata,
 
     C_ADD(fek, fk, fnkc);
     C_SUB(tmp, fk, fnkc);
-    C_MUL(fok, tmp, st->super_twiddles[k - 1]);
-    C_ADD(st->tmpbuf[k], fek, fok);
-    C_SUB(st->tmpbuf[ncfft - k], fek, fok);
+    C_MUL(fok, tmp, cfg->super_twiddles[k - 1]);
+    C_ADD(cfg->tmpbuf[k], fek, fok);
+    C_SUB(cfg->tmpbuf[ncfft - k], fek, fok);
 #ifdef USE_SIMD
-    st->tmpbuf[ncfft - k].i *= _mm_set1_ps(-1.0);
+        st->tmpbuf[ncfft - k].i *= _mm_set1_ps(-1.0);
 #else
-    st->tmpbuf[ncfft - k].i *= -1;
+    cfg->tmpbuf[ncfft - k].i *= -1;
 #endif
   }
-  kiss_fft(st->substate, st->tmpbuf, (kiss_fft_cpx *)timedata);
+  kiss_fft(cfg->substate, cfg->tmpbuf, reinterpret_cast<kiss_fft_cpx *>(timedata));
 }
