@@ -15,14 +15,12 @@ You should have received a copy of the GNU General Public License
 along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
+#pragma once
 
-#ifndef CHANNELMAPS_H
-#define CHANNELMAPS_H
 #include "FreeSurroundDecoder.h"
 #include <map>
-#include <vector>
 
-const int grid_res = 21; // resolution of the lookup grid
+constexpr int grid_res = 21; // resolution of the lookup grid
 
 // channel allocation maps (per setup)
 typedef std::vector<std::vector<float *>> alloc_lut;
@@ -32,5 +30,3 @@ extern std::map<unsigned, std::vector<float>> chn_angle;
 extern std::map<unsigned, std::vector<float>> chn_xsf;
 extern std::map<unsigned, std::vector<float>> chn_ysf;
 extern std::map<unsigned, std::vector<channel_id>> chn_id;
-
-#endif
