@@ -1,10 +1,13 @@
-#ifndef KISS_FFT_H
-#define KISS_FFT_H
+#pragma once
 
-#include <math.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
+#include <cmath>
+#include <cstring>
+#include <cstdlib>
+#include <numbers>
+
+using std::abs;
+using std::sqrt;
+using std::numbers::pi;
 
 #ifdef __cplusplus
 extern "C" {
@@ -51,8 +54,8 @@ extern "C" {
 #endif
 
 typedef struct {
-  kiss_fft_scalar r;
-  kiss_fft_scalar i;
+ kiss_fft_scalar r;
+ kiss_fft_scalar i;
 } kiss_fft_cpx;
 
 typedef struct kiss_fft_state *kiss_fft_cfg;
@@ -60,9 +63,9 @@ typedef struct kiss_fft_state *kiss_fft_cfg;
 /*
  *  kiss_fft_alloc
  *
- *  Initialize a FFT (or IFFT) algorithm's cfg/state buffer.
+ *  Initialize an FFT (or IFFT) algorithm's cfg/state buffer.
  *
- *  typical usage:      kiss_fft_cfg mycfg=kiss_fft_alloc(1024,0,NULL,NULL);
+ *  typical usage: kiss_fft_cfg mycfg=kiss_fft_alloc(1024,0,NULL,NULL);
  *
  *  The return value from fft_alloc is a cfg buffer used internally
  *  by the fft routine or NULL.
@@ -76,13 +79,12 @@ typedef struct kiss_fft_state *kiss_fft_cfg;
  *      then the function places the cfg in mem and the size used in *lenmem
  *      and returns mem.
  *
- *  If lenmem is not NULL and ( mem is NULL or *lenmem is not large enough),
+ *  If lenmem is not NULL and (mem is NULL or *lenmem is not large enough),
  *      then the function returns NULL and places the minimum cfg
  *      buffer size in *lenmem.
  * */
 
-kiss_fft_cfg kiss_fft_alloc(int nfft, int inverse_fft, void *mem,
-                            size_t *lenmem);
+kiss_fft_cfg kiss_fft_alloc(int nfft, int inverse_fft, void *mem, size_t *lenmem);
 
 /*
  * kiss_fft(cfg,in_out_buf)
@@ -100,8 +102,7 @@ void kiss_fft(kiss_fft_cfg cfg, const kiss_fft_cpx *fin, kiss_fft_cpx *fout);
  A more generic version of the above function. It reads its input from every Nth
  sample.
  * */
-void kiss_fft_stride(kiss_fft_cfg cfg, const kiss_fft_cpx *fin,
-                     kiss_fft_cpx *fout, int fin_stride);
+void kiss_fft_stride(kiss_fft_cfg cfg, const kiss_fft_cpx *fin, kiss_fft_cpx *fout, int fin_stride);
 
 /* If kiss_fft_alloc allocated a buffer, it is one contiguous
    buffer and can be simply free()d when no longer needed*/
@@ -121,11 +122,8 @@ void kiss_fft_cleanup(void);
 int kiss_fft_next_fast_size(int n);
 
 /* for real ffts, we need an even size */
-#define kiss_fftr_next_fast_size_real(n)                                       \
-  (kiss_fft_next_fast_size(((n) + 1) >> 1) << 1)
+#define kiss_fftr_next_fast_size_real(n) (kiss_fft_next_fast_size(((n) + 1) >> 1) << 1)
 
 #ifdef __cplusplus
 }
-#endif
-
 #endif
